Migrated MainActivity code to presenter
