+++ app/src/main/java/com/vest10/peter/madklubandroid/detail_activity/di/DetailActivityDependenciesModule.kt
@@ -10,7 +10,7 @@ import dagger.Provides
  */
 @Module
 class DetailActivityDependenciesModule {
-    @Provides
-    @ActivityScope
-    fun providesDetailPresenter() = DetailPresenter()
+    //@Provides
+    //@ActivityScope
+    //fun providesDetailPresenter() = DetailPresenter()
 }
