+++ app/src/main/java/com/vest10/peter/madklubandroid/main_activity/MainActivity.kt
@@ -22,21 +22,17 @@ import kotlinx.android.synthetic.main.upcomming_dinnerclub_item.view.*
 import javax.inject.Inject
 import android.support.v7.widget.DividerItemDecoration
 import com.vest10.peter.madklubandroid.depenedency_injection.components.ConfigPersistentComponent
+import com.vest10.peter.madklubandroid.upcomming_dinnerslubs_list.UpcommingDinnerclubItem
 
 
 class MainActivity : BaseActivity<MainPresenter.MainView,MainPresenter>(), MainPresenter.MainView {
-
     override fun injectMembers(configPersistentComponent: ConfigPersistentComponent) {
         configPersistentComponent.mainActivityComponent().inject(this)
     }
 
-    @Inject
-    lateinit var networkService: NetworkService
-
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
-        Log.d("MadklubTest","We have presenter: ${presenter}")
         kitchen_list.apply {
             setHasFixedSize(true)
             adapter = UpcommingDinnerclubsAdapter({
@@ -91,71 +87,13 @@ class MainActivity : BaseActivity<MainPresenter.MainView,MainPresenter>(), MainP
             )
             addItemDecoration(mDividerItemDecoration)
         }
-        kitchen_list.setHasFixedSize(true)
+    }
+
+    override fun showDinnerclubs(dinnerclubs: List<UpcommingDinnerclubItem>) {
+        (kitchen_list.adapter as UpcommingDinnerclubsAdapter).addDinnerclubs(dinnerclubs)
     }
 
     override fun launchAuthenticatedNetworkRequests() {
-        val getDinnerclubs = networkService.query {
-            UpcommingDinnerclubsQuery.builder()
-                    .startDate("2017-09-22T12:00:00.000Z")
-                    .endDate("2017-12-22T12:00:00.000Z")
-                    .build()
-        }
-        //.delay(4,TimeUnit.SECONDS)
-        .map {
-            val me = it.data()?.me()
-            Pair<String,List<UpcommingDinnerclubsQuery.Dinnerclub>>(me?.id()!!,me.kitchen()?.dinnerclubs()!!)
-            //it.data()?.me()?.kitchen()?.dinnerclubs()
-        }.onErrorReturn({
-            Log.d("We had the error", it.localizedMessage)
-            Log.d("We had the error", "$it")
-            Pair("",emptyList<UpcommingDinnerclubsQuery.Dinnerclub>())
-        }).map {
-            pair ->
-            pair.second.map {
-                val id = it.id()
-                val isParticipating = it.participants()!!.fold(false){
-                    part,p -> part || (if(p.user()!!.id() == pair.first) !(p.cancelled()!!) else false)
-                }
-                if(pair.first == it.cook().id()){
-                    // Current user is cook
-                    //throw RuntimeException("Do we get the cook case")
-                    CookDinnerclubItem(
-                            id,
-                            it.cancelled(),
-                            it.shopping_complete(),
-                            it.at(),
-                            it.meal(),
-                            isParticipating,
-                            // Summing the participants
-                            it.participants()!!.fold(0){
-                                sum,p -> sum + if(!p.cancelled()!!) 1+p.guest_count()!! else 0
-                            }
-                    )
-                }else{
-                    // Current user is NOT cook
-                    RegularDinnerclubItem(
-                            id,
-                            it.cancelled(),
-                            it.shopping_complete(),
-                            it.at(),
-                            it.meal(),
-                            isParticipating,
-                            it.cook().display_name()!!)
-                }
-            }
-        }.subscribe ({
-            res ->
-            (kitchen_list.adapter as UpcommingDinnerclubsAdapter).addDinnerclubs(res)
-            Log.d("Madklub","successfully returned from logged in query")
-        },{
-            // TODO do errors based on their type
-            error -> when (error) {
-        // TODO inform user there is no connection
-            is ApolloNetworkException -> Log.d("Madklub","Network error, server probably down...")
-        }
-            Log.d("Madklub","We had an error")
-        })
-        disposables.add(getDinnerclubs)
+        presenter.getDinnerclubs()
     }
 }
